added comments, fixed a few errors and moved contollers to handlers.
