+++ handlers/postController.js
@@ -1,5 +1,6 @@
-const models = require ('./models')
+const models = require('../models')
 
+//Grabs post and sends it to page
 module.exports.getPost = (req,res,next) => {
   models.Post.findByPk(postId,{
     include: [
