+++ routes/acct.js
@@ -1,6 +1,6 @@
 const express = require('express')
 const router = express.Router()
-const AcctController = require('../controllers/acctController')
+const AcctController = require('../handlers/acctController')
 
 router.get('/',AcctController.getYourPosts)
